Fixed watchOS sync, going to use SyncManager instead for data syncing between cloud and local
